requirements.txt
openai==0.27.1           
Pillow==9.6.0            
streamlit==1.25.0       
tqdm==4.66.0             
transformers==4.30.0     
